backend: use string with int64 format in api schema due to JS limitations, use int64 everywhere for steam_id values

+++ backend/internal/service/get_players.go
@@ -25,17 +25,17 @@ func (s *Service) GetPlayers(ctx context.Context, p gen.GetPlayersParams) (r *ge
 	}, nil
 }
 
-func (s *Service) getPlayers(ctx context.Context, playerIDs []int, withRecruitmentStatus bool) ([]gen.Player, error) {
+func (s *Service) getPlayers(ctx context.Context, playerIDs []int64, withRecruitmentStatus bool) ([]gen.Player, error) {
 	var players []gen.Player
 
 	for _, playerID := range playerIDs {
-		player, err := s.cache.GetPlayer(ctx, cache.LeagueETF2L, int64(playerID))
+		player, err := s.cache.GetPlayer(ctx, cache.LeagueETF2L, playerID)
 		switch {
 		case errors.Is(err, redis.Nil):
 			etf2lPlayer, err := s.etf2l.GetPlayer(ctx, playerID)
 			switch {
 			case errors.Is(err, etf2l.ErrPlayerNotFound):
-				if cacheErr := s.cache.SetPlayer(ctx, cache.LeagueETF2L, int64(playerID), cache.Player{DoesntExists: true}); cacheErr != nil {
+				if cacheErr := s.cache.SetPlayer(ctx, cache.LeagueETF2L, playerID, cache.Player{DoesntExists: true}); cacheErr != nil {
 					return nil, fmt.Errorf("failed to save unknown player to cache: %w", cacheErr)
 				}
 				continue
@@ -44,7 +44,7 @@ func (s *Service) getPlayers(ctx context.Context, playerIDs []int, withRecruitme
 			}
 
 			player = etf2lPlayer.ToCache()
-			if cacheErr := s.cache.SetPlayer(ctx, cache.LeagueETF2L, int64(playerID), player); cacheErr != nil {
+			if cacheErr := s.cache.SetPlayer(ctx, cache.LeagueETF2L, playerID, player); cacheErr != nil {
 				return nil, fmt.Errorf("failed to save player to cache: %w", cacheErr)
 			}
 		case err != nil:
@@ -72,7 +72,7 @@ func (s *Service) getPlayers(ctx context.Context, playerIDs []int, withRecruitme
 		}
 
 		if withRecruitmentStatus {
-			apiPlayer.Recruitment, err = s.getRecruitmentStatusForPlayer(ctx, int64(playerID))
+			apiPlayer.Recruitment, err = s.getRecruitmentStatusForPlayer(ctx, playerID)
 			if err != nil {
 				return nil, err
 			}
